Fixing OpLoad validation for variable pointers.
OpLoad expects a pointer as an argument. If the SPIR-V code has the
"SPV_KHR_variable_pointers" extension, then the pointer may be the result
of any of the following:
OpSelect
OpPhi
OpFunctionCall
OpPtrAccessChain
OpCopyObject
OpLoad
OpConstantNull

The validator was not taking this into account and was assuming that the
result of OpSelect for instance may not be used as an argument of
OpLoad.